requirements/base.txt
# Base requirements for ProductInsights

# Flask and core extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0

# Authentication and security
Werkzeug==2.3.7
WTForms==3.0.1
PyJWT==2.8.0
cryptography==41.0.4

# Data validation and serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# HTTP and API
requests==2.31.0
httpx==0.24.1

# Background tasks
celery==5.3.1
redis==4.6.0

# Data processing and analysis
pandas==2.0.3
numpy==1.24.3
textblob==0.17.1
nltk==3.8.1

# Web scraping
selenium==4.11.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.0
browser-cookie3==0.19.1

# Social media APIs
tweepy==4.14.0
instagrapi==2.0.0

# Data visualization
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3
gunicorn==21.2.0
